port all the capture spec without getting them passing
